+++ src/net/FrameworkDialogs/OpenFile/OpenFileDialogWrapper.cs
@@ -9,8 +9,8 @@ namespace MvvmDialogs.FrameworkDialogs.OpenFile
     /// </summary>
     internal sealed class OpenFileDialogWrapper : IFrameworkDialog
     {
-        private readonly OpenFileDialogSettings settings;
-        private readonly OpenFileDialog openFileDialog;
+        private readonly OpenFileDialog dialog;
+        private readonly OpenFileDialogSettingsSync sync;
 
         /// <summary>
         /// Initializes a new instance of the <see cref="OpenFileDialogWrapper"/> class.
@@ -18,21 +18,11 @@ namespace MvvmDialogs.FrameworkDialogs.OpenFile
         /// <param name="settings">The settings for the open file dialog.</param>
         public OpenFileDialogWrapper(OpenFileDialogSettings settings)
         {
-            this.settings = settings ?? throw new ArgumentNullException(nameof(settings));
+            dialog = new OpenFileDialog();
+            sync = new OpenFileDialogSettingsSync(dialog, settings);
 
-            openFileDialog = new OpenFileDialog
-            {
-                AddExtension = settings.AddExtension,
-                CheckFileExists = settings.CheckFileExists,
-                CheckPathExists = settings.CheckPathExists,
-                DefaultExt = settings.DefaultExt,
-                FileName = settings.FileName,
-                Filter = settings.Filter,
-                FilterIndex = settings.FilterIndex,
-                InitialDirectory = settings.InitialDirectory,
-                Multiselect = settings.Multiselect,
-                Title = settings.Title
-            };
+            // Update dialog
+            sync.ToDialog();
         }
 
         /// <inheritdoc />
@@ -40,11 +30,10 @@ namespace MvvmDialogs.FrameworkDialogs.OpenFile
         {
             if (owner == null) throw new ArgumentNullException(nameof(owner));
 
-            bool? result = openFileDialog.ShowDialog(owner);
+            bool? result = dialog.ShowDialog(owner);
 
             // Update settings
-            settings.FileName = openFileDialog.FileName;
-            settings.FileNames = openFileDialog.FileNames;
+            sync.ToSettings();
 
             return result;
         }